+++ src/main/java/subway/dto/request/LineRequest.java
@@ -1,0 +1,24 @@
+package subway.dto.request;
+
+public class LineRequest {
+
+    private String name;
+    private String color;
+
+    public LineRequest() {
+    }
+
+    public LineRequest(String name, String color) {
+        this.name = name;
+        this.color = color;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+}
