+++ src/main/java/subway/dto/request/StationRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.dto.request;
 
 public class StationRequest {
