+++ src/main/java/subway/ui/SectionController.java
@@ -1,18 +1,13 @@
 package subway.ui;
 
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import subway.application.SectionService;
+import subway.dto.request.SectionRequest;
+import subway.dto.response.SectionResponse;
+
 import java.net.URI;
 import java.sql.SQLException;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import subway.application.SectionService;
-import subway.dto.SectionRequest;
-import subway.dto.SectionResponse;
 
 @RestController
 public class SectionController {
