+++ src/main/java/subway/ui/StationController.java
@@ -1,21 +1,14 @@
 package subway.ui;
 
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import subway.application.StationService;
+import subway.dto.request.StationRequest;
+import subway.dto.response.StationResponse;
+
 import java.net.URI;
 import java.sql.SQLException;
 import java.util.List;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import subway.application.StationService;
-import subway.dto.StationRequest;
-import subway.dto.StationResponse;
 
 @RestController
 @RequestMapping("/stations")
@@ -46,7 +39,7 @@ public class StationController {
 
     @PutMapping("/{id}")
     public ResponseEntity<Void> updateStation(@PathVariable Long id,
-            @RequestBody StationRequest stationRequest) {
+                                              @RequestBody StationRequest stationRequest) {
         stationService.updateStation(id, stationRequest);
         return ResponseEntity.ok().build();
     }
