+++ src/test/java/subway/application/SectionServiceTest.java
@@ -1,16 +1,16 @@
 package subway.application;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
-
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 import subway.dao.SectionDao;
-import subway.dto.SectionRequest;
-import subway.dto.SectionResponse;
+import subway.dto.request.SectionRequest;
+import subway.dto.response.SectionResponse;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
 
 @SpringBootTest
 @Transactional
